Changed the package name from test.nice.testproject -> uk.co.jordanterry.espressoexamples #10
