+++ app/src/androidTest/java/uk/co/jordanterry/espressoexamples/AllTestsSuite.java
@@ -1,0 +1,21 @@
+package uk.co.jordanterry.espressoexamples;
+
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+
+/**
+ * Runs all instrumentation tests from one place
+ */
+@RunWith (Suite.class)
+@Suite.SuiteClasses ({
+		CustomIntentTests.class,
+		DeviceRotationTests.class,
+		HierarchyActivityTests.class,
+		MainActivityTests.class,
+		NavigationDrawerActivityTests.class,
+		PositionActivityTests.class,
+		SwitchActivityTests.class,
+		ViewPagerActivityTests.class
+})
+public class AllTestsSuite {
+}
