+++ app/src/main/AndroidManifest.xml
@@ -8,23 +8,23 @@
         android:label="@string/app_name"
         android:theme="@style/Theme.EspressoTestTheme">
         <activity
-            android:name=".activities.MainActivity"
+            android:name="uk.co.jordanterry.espressoexamples.activities.MainActivity"
             android:label="@string/app_name">
         </activity>
         <activity
-            android:name=".activities.SwitchActivity"
+            android:name="uk.co.jordanterry.espressoexamples.activities.SwitchActivity"
             android:label="@string/app_name">
         </activity>
         <activity
-            android:name=".activities.NavigationDrawerActivity"
+            android:name="uk.co.jordanterry.espressoexamples.activities.NavigationDrawerActivity"
             android:label="@string/app_name">
         </activity>
         <activity
-            android:name=".activities.PositionActivity"
+            android:name="uk.co.jordanterry.espressoexamples.activities.PositionActivity"
             android:label="@string/app_name">
         </activity>
         <activity
-            android:name=".activities.ViewPagerActivity"
+            android:name="uk.co.jordanterry.espressoexamples.activities.ViewPagerActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
@@ -33,11 +33,11 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".activities.HierarchyActivity"
+            android:name="uk.co.jordanterry.espressoexamples.activities.HierarchyActivity"
             android:label="@string/app_name">
         </activity>
         <activity
-            android:name=".activities.CustomIntentActivity"
+            android:name="uk.co.jordanterry.espressoexamples.activities.CustomIntentActivity"
             android:label="@string/app_name"/>
 
     </application>
