--- app/src/main/java/test/nice/testproject/activities/NavigationDrawerActivity.java
@@ -1,89 +1,0 @@
-package test.nice.testproject.activities;
-
-import android.content.res.Configuration;
-import android.os.Bundle;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.Gravity;
-import android.view.MenuItem;
-import android.view.View;
-import android.widget.AdapterView;
-import android.widget.ArrayAdapter;
-import android.widget.ListView;
-
-import test.nice.testproject.R;
-import test.nice.testproject.fragments.NavigationFragment;
-import test.nice.testproject.fragments.TestFragment;
-
-
-public class NavigationDrawerActivity extends AppCompatActivity {
-
-	private DrawerLayout mDrawerLayout;
-	private ActionBarDrawerToggle mDrawerToggle;
-	private String[] mMenuItems;
-
-	@Override
-	protected void onCreate(Bundle savedInstanceState) {
-		super.onCreate(savedInstanceState);
-		setContentView(R.layout.activity_navigation_drawer);
-		Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
-		setSupportActionBar(toolbar);
-		mDrawerLayout = (DrawerLayout) findViewById(R.id.my_drawer_layout);
-		mDrawerToggle = new ActionBarDrawerToggle(this, mDrawerLayout, toolbar, R.string.app_name, R.string.app_name);
-		mDrawerLayout.setDrawerListener(mDrawerToggle);
-		mMenuItems = new String[] {"Menu One", "Menu Two", "Menu Three", "Menu Four"};
-		ArrayAdapter<String> itemsAdapter = new ArrayAdapter<String>(this, android.R.layout.simple_list_item_1, mMenuItems);
-
-		ListView navigationList = (ListView) findViewById(R.id.navigation_menu);
-		navigationList.setOnItemClickListener(new NavigationDrawerClickListener());
-		navigationList.setAdapter(itemsAdapter);
-
-		if (savedInstanceState == null) {
-			getFragmentManager().beginTransaction().add(R.id.container, TestFragment.newInstance()).commit();
-		}
-	}
-
-	@Override
-	public boolean onOptionsItemSelected(MenuItem item) {
-		if (mDrawerToggle.onOptionsItemSelected(item)) {
-			return true;
-		}
-		return super.onOptionsItemSelected(item);
-	}
-
-
-	@Override
-	protected void onPostCreate(Bundle savedInstanceState) {
-		super.onPostCreate(savedInstanceState);
-		mDrawerToggle.syncState();
-	}
-
-	@Override
-	public void onConfigurationChanged(Configuration newConfig) {
-		super.onConfigurationChanged(newConfig);
-		mDrawerToggle.onConfigurationChanged(newConfig);
-	}
-
-	@Override
-	public void onBackPressed() {
-		if (mDrawerLayout.isDrawerOpen(Gravity.START | Gravity.LEFT)) {
-			mDrawerLayout.closeDrawers();
-			return;
-		}
-		super.onBackPressed();
-	}
-
-	public class NavigationDrawerClickListener implements ListView.OnItemClickListener {
-
-		@Override
-		public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-			getFragmentManager().beginTransaction().add(R.id.container,
-					NavigationFragment.newInstance(position, mMenuItems[position]), NavigationFragment.TAG)
-					.addToBackStack(NavigationFragment.TAG).commit();
-			mDrawerLayout.closeDrawers();
-		}
-	}
-
-}
