--- app/src/main/java/test/nice/testproject/activities/PositionActivity.java
@@ -1,24 +1,0 @@
-package test.nice.testproject.activities;
-
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-
-import test.nice.testproject.R;
-import test.nice.testproject.fragments.PositionFragment;
-
-
-public class PositionActivity extends AppCompatActivity {
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-        if (savedInstanceState == null) {
-            getFragmentManager().beginTransaction().add(R.id.container, PositionFragment.newInstance()).commit();
-        }
-    }
-
-
-}
