+++ app/src/main/java/uk/co/jordanterry/espressoexamples/activities/MainActivity.java
@@ -1,0 +1,33 @@
+package uk.co.jordanterry.espressoexamples.activities;
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
+import android.view.Menu;
+
+import test.nice.testproject.R;
+import uk.co.jordanterry.espressoexamples.fragments.TestFragment;
+
+
+public class MainActivity extends AppCompatActivity {
+
+	@Override
+	protected void onCreate(Bundle savedInstanceState) {
+		super.onCreate(savedInstanceState);
+		setContentView(R.layout.activity_main);
+		Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+		setSupportActionBar(toolbar);
+
+		if (savedInstanceState == null) {
+			getFragmentManager().beginTransaction().add(R.id.container, TestFragment.newInstance()).commit();
+		}
+	}
+
+
+	@Override
+	public boolean onCreateOptionsMenu(Menu menu) {
+		getMenuInflater().inflate(R.menu.menu_main, menu);
+		return true;
+	}
+
+}
