+++ app/src/main/java/uk/co/jordanterry/espressoexamples/adapters/PagerAdapter.java
@@ -1,0 +1,34 @@
+package uk.co.jordanterry.espressoexamples.adapters;
+
+import android.app.Fragment;
+import android.app.FragmentManager;
+import android.support.v13.app.FragmentStatePagerAdapter;
+
+import uk.co.jordanterry.espressoexamples.fragments.PagerFragment;
+
+/**
+ * Created by jordan_terry on 05/02/2015.
+ */
+public class PagerAdapter extends FragmentStatePagerAdapter {
+    public String[] mStrings;
+
+
+    public PagerAdapter(FragmentManager fm) {
+        super(fm);
+    }
+
+    public PagerAdapter(FragmentManager fm, String[] strings) {
+        super(fm);
+        mStrings = strings;
+    }
+
+    @Override
+    public Fragment getItem(int position) {
+        return new PagerFragment().newInstance(mStrings[position]);
+    }
+
+    @Override
+    public int getCount() {
+        return mStrings.length;
+    }
+}
